Introduce first-class strategies and metrics to rewrite_serach
As a consequence, the quality of the core code has been greatly improved. A great deal of ugly pieces have been able to removed, which were essentially present because we pesudo-supported extensions with other selection strategies (now we do this in a honest way).

Essentially every line of the rewrite_search engine has been touched.